+++ docs/scripts/api-events.py
@@ -75,6 +75,7 @@ with outfile.open("w") as f, contextlib.redirect_stdout(f):
         "",
         [
             addonmanager.LoadHook,
+            hooks.StartupHook,
             hooks.RunningHook,
             hooks.ConfigureHook,
             hooks.DoneHook,