+++ mitmproxy/addons/proxyserver.py
@@ -127,10 +127,14 @@ class Proxyserver(ServerManager):
             """Set the local IP address that mitmproxy should use when connecting to upstream servers.""",
         )
 
-    async def running(self):
+    async def startup(self):
+        errs = await self.setup_servers()
+        if errs:
+            errs_str = "\n".join(str(err) for err in errs)
+            raise exceptions.OptionsError(f"The following server instances failed to start:\n\n{errs_str}")
+
+    def running(self):
         self.is_running = True
-        # TODO: Do this before running()
-        await self.setup_servers()
 
     def configure(self, updated):
         if "stream_large_bodies" in updated:
@@ -195,8 +199,8 @@ class Proxyserver(ServerManager):
             if self.is_running:
                 asyncio.create_task(self.setup_servers())
 
-    async def setup_servers(self) -> bool:
-        all_ok = True
+    async def setup_servers(self) -> list[Exception]:
+        errs = []
         async with self._lock:
             new_servers: dict[str, ServerInstance] = dict.fromkeys(ctx.options.mode)  # type: ignore
             if not ctx.options.server:
@@ -209,7 +213,7 @@ class Proxyserver(ServerManager):
             ]
             for ret in await asyncio.gather(*shutdown_tasks, return_exceptions=True):
                 if ret:
-                    all_ok = False
+                    errs.append(ret)
                     ctx.log.error(str(ret))
 
             new_instances: list[ServerInstance] = []
@@ -223,11 +227,11 @@ class Proxyserver(ServerManager):
 
             for ret in await asyncio.gather(*[m.start() for m in new_instances], return_exceptions=True):
                 if ret:
-                    all_ok = False
+                    errs.append(ret)
                     ctx.log.error(str(ret))
 
             self.servers = new_servers
-            return all_ok
+            return errs
 
     def listen_addrs(self) -> list[Address]:
         return [